Comprehensive code review fixes: security, efficiency, and consistency
Security fixes:
- Replace dangerouslySetInnerHTML with SafeHtml component using DOMPurify
- Fix SQL injection in dashboard.php using $wpdb->prepare()
- Add output escaping to theme files (functions.php, footer.php, single.php)
- Add strict type casting (int) to PHP permission checks
- Fix loose comparison (== to ===) in theme sanitize_checkbox

Efficiency fixes:
- Add staleTime to 12+ React Query hooks to prevent unnecessary refetches
- Fix useMemo dependency array in FileGrid.tsx (add resize listener)
- Fix LazyView function recreation in App.tsx using useMemo

Consistency fixes:
- Update readme.txt versions to match plugin headers (3.6.4, 2.1.0, 2.2.0)
- Move Family types from client.ts to types/index.ts
- Add chatSearch to queryKeys object for consistency
- Update portal package.json version to 2.1.0 and add type-check script

Build status: ✓ Tests passing (45/45), ✓ Lint clean, ✓ Build successful

+++ france-relocation-member-tools/includes/class-framt-portal-api.php
@@ -1193,7 +1193,7 @@ class FRAMT_Portal_API {
             );
         }
 
-        if ( $project->user_id !== get_current_user_id() && ! current_user_can( 'manage_options' ) ) {
+        if ( (int) $project->user_id !== (int) get_current_user_id() && ! current_user_can( 'manage_options' ) ) {
             return new WP_Error(
                 'rest_forbidden',
                 'You do not have permission to access this project.',
@@ -1227,7 +1227,7 @@ class FRAMT_Portal_API {
             );
         }
 
-        if ( $task->user_id !== get_current_user_id() && ! current_user_can( 'manage_options' ) ) {
+        if ( (int) $task->user_id !== (int) get_current_user_id() && ! current_user_can( 'manage_options' ) ) {
             return new WP_Error(
                 'rest_forbidden',
                 'You do not have permission to access this task.',
@@ -1261,7 +1261,7 @@ class FRAMT_Portal_API {
             );
         }
 
-        if ( $task->user_id !== get_current_user_id() && ! current_user_can( 'manage_options' ) ) {
+        if ( (int) $task->user_id !== (int) get_current_user_id() && ! current_user_can( 'manage_options' ) ) {
             return new WP_Error(
                 'rest_forbidden',
                 'You do not have permission to access this task.',
